Moved TxMA related interfaces to relevant location
