+++ src/utils/TxmaEvent.ts
@@ -1,4 +1,4 @@
-import { TxMAVerifiedCredential } from "./IVeriCredential";
+import { CredentialSubject } from "./IVeriCredential";
 import { ISessionItem } from "../models/ISessionItem";
 
 export type TxmaEventName =
@@ -23,6 +23,18 @@ export interface Evidence {
 
 export interface Extensions {
 	"evidence": Evidence;
+}
+
+export interface TxMACredentialSubject extends CredentialSubject {
+	device_information?: {
+		encoded: string;
+	};
+}
+
+export interface TxMAVerifiedCredential {
+	"@context": string[];
+	type: string[];
+	credentialSubject: TxMACredentialSubject;
 }
 
 export interface BaseTxmaEvent {
